requirements.txt
# CrewAI and AI orchestration
crewai>=0.83.0
crewai-tools>=0.12.0

# LLM inference
llama-cpp-python>=0.2.77
huggingface-hub>=0.20.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0.1
requests>=2.31.0
aiohttp>=3.9.0

# Logging and monitoring
structlog>=23.2.0
rich>=13.7.0
psutil>=5.9.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
isort>=5.13.0
mypy>=1.8.0

# Optional: For advanced OCR preprocessing (if needed)
opencv-python>=4.8.0
Pillow>=10.2.0
python-dotenv
flask>=2.3.0
